// apps/web-v4/src/app/models/certification.model.ts
export interface ICertificationRequestItem {
  record_type: TCertificationRequestType;
  user?: {
    name: string;
    email: string;
  };
  currently_with: {
    name: string;
    email: string;
  };
  raised_on?: Date;
  status?: string;
  certification: string;
  certification_name: string;
  booking_type?: string;
  country?: string;
  location?: string;
  date?: Date;
  slot?: string;
  slotno?: number;
  user_contact?: string;
  icfdid?: number;
  sino?: number;
  ecdp_id?: number;
  amount?: number;
  currency?: string;
  exam_date?: Date;
  document?: {
    document_url: string;
  }[];
  document_url?: string;
  resultType?: string;
  result?: string;
  upload_id?: number;
}

export interface ICertificationApproverAction {
  status: string;
  reason?: string;
  upload_id?: number;
  user?: string;
}

export interface ICertificationSubmission {
  certification?: string;
  certification_name: string;
  certification_type: string;
  examDate: Date;
  result_type: string;
  result: string;
  status?: string;
}

export interface ICertificationApproverData {
  certificationId?: string;
  icfdId?: number;
  ecdpId?: number;
  sino?: number;
  status?: 'Approved' | 'Rejected';
  reason?: string;
  uploadId?: number;
  user?: {
    name: string;
    email: string;
  };
}

export interface ICertificationSendResponse {
  res_code: number;
  res_message: string;
}

export interface IResultUploadForm {
  result_type: string;
  result: string;
  exam_date: number | Date;
  fileName: string;
  file: File;
}

export interface ICertificationDialogData {
  approvalItem: ICertificationRequestItem;
  action: TCertificationRequestType;
  actionType: TCertificationActionType;
}

export interface IUserActionDialogData {
  approvalItem: ICertificationRequestItem;
  actionType: TCertificationUserActionType;
}

export interface ICertificationDialogResult {
  action: TCertificationRequestType;
  result: boolean;
  data: ICertificationApproverData;
}
// ...
export type TCertificationRequestType = 'result_verification' | 'budget_approval' | 'proctor_approval' | 'acc' | 'all';
export type TCertificationActionType = 'accept' | 'decline' | 'submit';